Remove compilation warnings + clearer makefile output

+++ 3rdParty/LightPcapNg/LightPcapNg/src/light_manipulate.c
@@ -466,7 +466,7 @@ int light_ip_flow(light_pcapng *sectionp, light_pcapng **flows, size_t *flow_cou
 iterate:
 		progress++;
 		if (progress % 10000 == 0) {
-			printf("Flow extraction progress: %.2lf [%zu / %zu]\n", (double)progress / limit * 100.0, progress, limit);
+			printf("Flow extraction progress: %.2lf [%d / %d]\n", (double)progress / limit * 100.0, (int)progress, (int)limit);
 		}
 		current_block = current_block->next_block;
 	}
